requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-logging
google-auth-httplib2 
google-auth-oauthlib
google-api-python-client
st-files-connection
google.generativeai
streamlit_js_eval